Implemented `/api/{{version}}/servers` mock endpoint
This had a LOT of dependencies, so this also includes the following
mock endpoints:

* /api/{{version}}/phys_locations
* /api/{{version}}/regions
* /api/{{version}}/divisions
* /api/{{version}}/types
* /api/{{version}}/statuses

This commit also contains a refactor to split up functionality for
related concepts into isolated files.
