+++ traffic_ops/client_tests/parameters.go
@@ -1,0 +1,822 @@
+package main
+
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+import "encoding/json"
+import "net/http"
+
+import "github.com/apache/trafficcontrol/lib/go-tc"
+import "github.com/apache/trafficcontrol/traffic_ops/traffic_ops_golang/api"
+
+var PARAMETERS = []tc.Parameter{
+	tc.Parameter{
+		ConfigFile:  "global",
+		ID:          1,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "tm.url",
+		Profiles:    json.RawMessage(`["` + GLOBAL_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "https://localhost:443/",
+	},
+	tc.Parameter{
+		ConfigFile:  "global",
+		ID:          2,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "tm.instance_name",
+		Profiles:    json.RawMessage(`["` + GLOBAL_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "Mock Traffic Ops CDN",
+	},
+	tc.Parameter{
+		ConfigFile:  "global",
+		ID:          3,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "tm.toolname",
+		Profiles:    json.RawMessage(`["` + GLOBAL_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "Mock Traffic Ops",
+	},
+	tc.Parameter{
+		ConfigFile:  "CRConfig.json",
+		ID:          4,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "geolocation.polling.url",
+		Profiles:    json.RawMessage(`["` + GLOBAL_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "https://localhost/mock/geo/database.dat",
+	},
+	tc.Parameter{
+		ConfigFile:  "CRConfig.json",
+		ID:          5,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "geolocation6.polling.url",
+		Profiles:    json.RawMessage(`["` + GLOBAL_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "https://localhost/mock/geo/database.dat",
+	},
+	tc.Parameter{
+		ConfigFile:  "global",
+		ID:          6,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "tm.logourl",
+		Profiles:    json.RawMessage(`["` + GLOBAL_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "/logo.svg",
+	},
+	tc.Parameter{
+		ConfigFile:  "global",
+		ID:          7,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "default_geo_miss_latitude",
+		Profiles:    json.RawMessage(`["` + GLOBAL_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "0",
+	},
+	tc.Parameter{
+		ConfigFile:  "global",
+		ID:          8,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "default_geo_miss_longitude",
+		Profiles:    json.RawMessage(`["` + GLOBAL_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "0",
+	},
+	tc.Parameter{
+		ConfigFile:  "regex_revalidate.config",
+		ID:          9,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "maxRevalDurationDays",
+		Profiles:    json.RawMessage(`["` + GLOBAL_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "90",
+	},
+	tc.Parameter{
+		ConfigFile:  "global",
+		ID:          10,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "use_reval_pending",
+		Profiles:    json.RawMessage(`["` + GLOBAL_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "1",
+	},
+	tc.Parameter{
+		ConfigFile:  "global",
+		ID:          11,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "use_tenancy",
+		Profiles:    json.RawMessage(`["` + GLOBAL_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "1",
+	},
+	tc.Parameter{
+		ConfigFile:  "records.config",
+		ID:          12,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "CONFIG proxy.config.http.server_ports",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "STRING 80 80:ipv6 443:proto=http:ssl 443:ipv6:proto=http:ssl",
+	},
+	tc.Parameter{
+		ConfigFile:  "records.config",
+		ID:          13,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "CONFIG proxy.config.reverse_proxy.enabled",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "INT 1",
+	},
+	tc.Parameter{
+		ConfigFile:  "records.config",
+		ID:          14,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "CONFIG proxy.config.ssl.CA.cert.path",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "STRING /etc/trafficserver/ssl",
+	},
+	tc.Parameter{
+		ConfigFile:  "records.config",
+		ID:          15,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "CONFIG proxy.config.ssl.client.CA.cert.path",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "STRING /etc/trafficserver/ssl",
+	},
+	tc.Parameter{
+		ConfigFile:  "records.config",
+		ID:          16,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "CONFIG proxy.config.ssl.client.cert.path",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "STRING /etc/trafficserver/ssl",
+	},
+	tc.Parameter{
+		ConfigFile:  "records.config",
+		ID:          17,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "CONFIG proxy.config.ssl.client.private_key.path",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "STRING /etc/trafficserver/ssl",
+	},
+	tc.Parameter{
+		ConfigFile:  "records.config",
+		ID:          18,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "CONFIG proxy.config.ssl.oscp.enabled",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "INT 1",
+	},
+	tc.Parameter{
+		ConfigFile:  "records.config",
+		ID:          19,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "CONFIG proxy.config.ssl.server.cert.path",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "STRING /etc/trafficserver/ssl",
+	},
+	tc.Parameter{
+		ConfigFile:  "records.config",
+		ID:          20,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "CONFIG proxy.config.ssl.server.ticket_key.filename",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "STRING NULL",
+	},
+	tc.Parameter{
+		ConfigFile:  "records.config",
+		ID:          21,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "CONFIG proxy.config.ssl.server.private_key.path",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "STRING /etc/trafficserver/ssl",
+	},
+	tc.Parameter{
+		ConfigFile:  "ssl_multicert.config",
+		ID:          22,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "location",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "/etc/trafficserver",
+	},
+	tc.Parameter{
+		ConfigFile:  "records.config",
+		ID:          23,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "CONFIG proxy.config.reverse_proxy.enabled",
+		Profiles:    json.RawMessage(`["` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "INT 0",
+	},
+	tc.Parameter{
+		ConfigFile:  "astats.config",
+		ID:          24,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "location",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "/etc/trafficserver",
+	},
+	tc.Parameter{
+		ConfigFile:  "astats.config",
+		ID:          25,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "allow_ip",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "0.0.0.0/0",
+	},
+	tc.Parameter{
+		ConfigFile:  "astats.config",
+		ID:          26,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "allow_ip6",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "::/0",
+	},
+	tc.Parameter{
+		ConfigFile:  "astats.config",
+		ID:          27,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "path",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "_astats",
+	},
+	tc.Parameter{
+		ConfigFile:  "astats.config",
+		ID:          28,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "record_types",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "122",
+	},
+	tc.Parameter{
+		ConfigFile:  "cache.config",
+		ID:          29,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "location",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "/etc/trafficserver",
+	},
+	tc.Parameter{
+		ConfigFile:  "chkconfig",
+		ID:          30,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "trafficserver",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "0:off\t1:off\t2:on\t3:on\t4:on\t5:on\t6:off",
+	},
+	tc.Parameter{
+		ConfigFile:  "hosting.config",
+		ID:          31,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "location",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "/etc/trafficserver",
+	},
+	tc.Parameter{
+		ConfigFile:  "ip_allow.config",
+		ID:          32,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "location",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "/etc/trafficserver",
+	},
+	tc.Parameter{
+		ConfigFile:  "parent.config",
+		ID:          33,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "location",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "/etc/trafficserver",
+	},
+	tc.Parameter{
+		ConfigFile:  "plugin.config",
+		ID:          34,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "location",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "/etc/trafficserver",
+	},
+	tc.Parameter{
+		ConfigFile:  "plugin.config",
+		ID:          35,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "astats_over_http.so",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "",
+	},
+	tc.Parameter{
+		ConfigFile:  "plugin.config",
+		ID:          36,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "regex_revalidate.so",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "--config regex_revalidate.config",
+	},
+	tc.Parameter{
+		ConfigFile:  "rascal.properties",
+		ID:          37,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "health.connection.timeout",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "2000",
+	},
+	tc.Parameter{
+		ConfigFile:  "rascal.properties",
+		ID:          38,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "health.polling.url",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "http://${hostname}/_astats?application=&inf.name=${interface_name}",
+	},
+	tc.Parameter{
+		ConfigFile:  "rascal.properties",
+		ID:          39,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "health.threshold.availableBandwidthInKbps",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       ">1750000",
+	},
+	tc.Parameter{
+		ConfigFile:  "rascal.properties",
+		ID:          40,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "health.threshold.loadavg",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "25.0",
+	},
+	tc.Parameter{
+		ConfigFile:  "rascal.properties",
+		ID:          41,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "health.threshold.queryTime",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "1000",
+	},
+	tc.Parameter{
+		ConfigFile:  "rascal.properties",
+		ID:          42,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "history.count",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "30",
+	},
+	tc.Parameter{
+		ConfigFile:  "records.config",
+		ID:          43,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "location",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "/etc/trafficserver",
+	},
+	tc.Parameter{
+		ConfigFile:  "records.config",
+		ID:          44,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "CONFIG proxy.config.admin.user_id",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "STRING ats",
+	},
+	tc.Parameter{
+		ConfigFile:  "records.config",
+		ID:          45,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "CONFIG proxy.config.body_factory.template_sets_dir",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "STRING /etc/trafficserver/body_factory",
+	},
+	tc.Parameter{
+		ConfigFile:  "records.config",
+		ID:          46,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "CONFIG proxy.config.config_dir",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "STRING /etc/trafficserver",
+	},
+	tc.Parameter{
+		ConfigFile:  "records.config",
+		ID:          47,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "CONFIG proxy.config.diags.debug.enabled",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "INT 1",
+	},
+	tc.Parameter{
+		ConfigFile:  "records.config",
+		ID:          48,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "CONFIG proxy.config.dns.round_robin_nameservers",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "INT 0",
+	},
+	tc.Parameter{
+		ConfigFile:  "records.config",
+		ID:          49,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "CONFIG proxy.config.exec_thread.autoconfig",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "INT 0",
+	},
+	tc.Parameter{
+		ConfigFile:  "records.config",
+		ID:          50,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "CONFIG proxy.config.http.cache.required_headers",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "INT 0",
+	},
+	tc.Parameter{
+		ConfigFile:  "records.config",
+		ID:          51,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "CONFIG proxy.config.http.enable_http_stats",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "INT 1",
+	},
+	tc.Parameter{
+		ConfigFile:  "records.config",
+		ID:          52,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "CONFIG proxy.config.http.insert_response_via_str",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "INT 3",
+	},
+	tc.Parameter{
+		ConfigFile:  "records.config",
+		ID:          53,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "CONFIG proxy.config.http.server_ports",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "STRING 80 80:ipv6",
+	},
+	tc.Parameter{
+		ConfigFile:  "records.config",
+		ID:          54,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "CONFIG proxy.config.http.slow.log.threshold",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "INT 10000",
+	},
+	tc.Parameter{
+		ConfigFile:  "records.config",
+		ID:          55,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "CONFIG proxy.config.http.transaction_active_timeout_in",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "INT 0",
+	},
+	tc.Parameter{
+		ConfigFile:  "records.config",
+		ID:          56,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "CONFIG proxy.config.log.logfile_dir",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "STRING /var/log/trafficserver",
+	},
+	tc.Parameter{
+		ConfigFile:  "records.config",
+		ID:          57,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "CONFIG proxy.config.http.parent_proxy.retry_time",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "INT 60",
+	},
+	tc.Parameter{
+		ConfigFile:  "records.config",
+		ID:          58,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "CONFIG proxy.config.http.parent_proxy_routing_enable",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "INT 1",
+	},
+	tc.Parameter{
+		ConfigFile:  "records.config",
+		ID:          59,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "CONFIG proxy.config.proxy_name",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "STRING __HOSTNAME__",
+	},
+	tc.Parameter{
+		ConfigFile:  "records.config",
+		ID:          60,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "CONFIG proxy.config.url_remap.remap_required",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "INT 0",
+	},
+	tc.Parameter{
+		ConfigFile:  "regex_revalidate.config",
+		ID:          61,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "location",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "/etc/trafficserver",
+	},
+	tc.Parameter{
+		ConfigFile:  "remap.config",
+		ID:          62,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "location",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "/etc/trafficserver",
+	},
+	tc.Parameter{
+		ConfigFile:  "set_dscp_0.config",
+		ID:          63,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "location",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "/etc/trafficserver/dscp",
+	},
+	tc.Parameter{
+		ConfigFile:  "set_dscp_8.config",
+		ID:          64,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "location",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "/etc/trafficserver/dscp",
+	},
+	tc.Parameter{
+		ConfigFile:  "set_dscp_10.config",
+		ID:          65,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "location",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "/etc/trafficserver/dscp",
+	},
+	tc.Parameter{
+		ConfigFile:  "set_dscp_12.config",
+		ID:          66,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "location",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "/etc/trafficserver/dscp",
+	},
+	tc.Parameter{
+		ConfigFile:  "set_dscp_14.config",
+		ID:          67,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "location",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "/etc/trafficserver/dscp",
+	},
+	tc.Parameter{
+		ConfigFile:  "set_dscp_16.config",
+		ID:          68,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "location",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "/etc/trafficserver/dscp",
+	},
+	tc.Parameter{
+		ConfigFile:  "set_dscp_18.config",
+		ID:          69,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "location",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "/etc/trafficserver/dscp",
+	},
+	tc.Parameter{
+		ConfigFile:  "set_dscp_22.config",
+		ID:          70,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "location",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "/etc/trafficserver/dscp",
+	},
+	tc.Parameter{
+		ConfigFile:  "set_dscp_24.config",
+		ID:          71,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "location",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "/etc/trafficserver/dscp",
+	},
+	tc.Parameter{
+		ConfigFile:  "set_dscp_26.config",
+		ID:          72,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "location",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "/etc/trafficserver/dscp",
+	},
+	tc.Parameter{
+		ConfigFile:  "set_dscp_28.config",
+		ID:          73,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "location",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "/etc/trafficserver/dscp",
+	},
+	tc.Parameter{
+		ConfigFile:  "set_dscp_30.config",
+		ID:          74,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "location",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "/etc/trafficserver/dscp",
+	},
+	tc.Parameter{
+		ConfigFile:  "set_dscp_32.config",
+		ID:          75,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "location",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "/etc/trafficserver/dscp",
+	},
+	tc.Parameter{
+		ConfigFile:  "set_dscp_34.config",
+		ID:          76,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "location",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "/etc/trafficserver/dscp",
+	},
+	tc.Parameter{
+		ConfigFile:  "set_dscp_36.config",
+		ID:          77,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "location",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "/etc/trafficserver/dscp",
+	},
+	tc.Parameter{
+		ConfigFile:  "set_dscp_37.config",
+		ID:          78,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "location",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "/etc/trafficserver/dscp",
+	},
+	tc.Parameter{
+		ConfigFile:  "set_dscp_38.config",
+		ID:          79,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "location",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "/etc/trafficserver/dscp",
+	},
+	tc.Parameter{
+		ConfigFile:  "set_dscp_40.config",
+		ID:          80,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "location",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "/etc/trafficserver/dscp",
+	},
+	tc.Parameter{
+		ConfigFile:  "set_dscp_48.config",
+		ID:          81,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "location",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "/etc/trafficserver/dscp",
+	},
+	tc.Parameter{
+		ConfigFile:  "set_dscp_56.config",
+		ID:          82,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "location",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "/etc/trafficserver/dscp",
+	},
+	tc.Parameter{
+		ConfigFile:  "storage.config",
+		ID:          83,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "location",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "/etc/trafficserver",
+	},
+	tc.Parameter{
+		ConfigFile:  "storage.config",
+		ID:          84,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "Disk_Volume",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "1",
+	},
+	tc.Parameter{
+		ConfigFile:  "storage.config",
+		ID:          85,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "Drive_Letters",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "cache",
+	},
+	tc.Parameter{
+		ConfigFile:  "storage.config",
+		ID:          86,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "Drive_Prefix",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "/var/trafficserver",
+	},
+	tc.Parameter{
+		ConfigFile:  "volume.config",
+		ID:          87,
+		LastUpdated: *CURRENT_TIME,
+		Name:        "location",
+		Profiles:    json.RawMessage(`["` + EDGE_PROFILE_NAME + `","` + MID_PROFILE_NAME + `"]`),
+		Secure:      false,
+		Value:       "/etc/trafficserver",
+	},
+}
+
+func parameters(w http.ResponseWriter, r *http.Request) {
+	common(w)
+	if (r.Method == http.MethodGet) {
+		api.WriteResp(w, r, PARAMETERS)
+	} else {
+		w.Header().Set("Allow", http.MethodGet)
+		w.WriteHeader(http.StatusNotImplemented)
+		w.Write([]byte(`{"alerts":[{"level":"error","text":"This method hasn't yet been implemented."}]}`))
+	}
+}
